updates in footer

+++ src/components/pacients/Pacients.jsx
@@ -1,61 +1,62 @@
-import React from "react";
+/* eslint-disable react/prop-types */
 import axios from "axios";
+import { STATUS } from "../constant";
 import Swal from "sweetalert2";
 import { Link } from "react-router-dom";
-import { STATUS } from "../constant";
-
-
 const Pacients = ({ pacients, getAPI }) => {
   const URL = import.meta.env.VITE_API_VETERINARIA;
+  const handleDelete = (id) => {
+    Swal.fire({
+      title: "Are you sure?",
 
-  const handleDelete = async (id) => {
-    try {
-      const res = await axios.delete(`${URL}/${id}`);
+      text: "You won't be able to revert this!",
 
-      console.log(res);
-      if (res.status === STATUS.STATUS_OK) {
-        Swal.fire("Deleted!", "Your product has been deleted", "success");
-        getAPI();
+      icon: "warning",
+
+      showCancelButton: true,
+
+      confirmButtonColor: "#3085d6",
+
+      cancelButtonColor: "#d33",
+
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const res = await axios.delete(`${URL}/${id}`);
+
+          console.log(res);
+          if (res.status === STATUS.STATUS_OK) {
+            Swal.fire("Deleted!", "Your product has been deleted", "success");
+            getAPI();
+          }
+        } catch (error) {
+          console.log(error);
+        }
       }
-    } catch (error) {
-      console.log(error);
-    }
+    });
   };
 
   return (
     <tr>
-      <td>
-      {pacients?._id}
-      </td>
-      <td>
-       {pacients?.ownerName}
-      </td>
-      <td>
-     {pacients?.email}
-      </td>
-      <td>
-   {pacients?.tel}
-      </td>
-      <td>
-       {pacients?.petName}
-      </td>
-      <td>
-      {pacients?.specie}
-      </td>
-      <td>
-       {pacients?.race}
-      </td>
+      <td>{pacients?._id}</td>
+      <td>{pacients?.ownerName}</td>
+      <td>{pacients?.email}</td>
+      <td>{pacients?.tel}</td>
+      <td>{pacients?.petName}</td>
+      <td>{pacients?.specie}</td>
+      <td>{pacients?.race}</td>
 
-      <td> 
-        <div className="d-flex flex-column align-items-center justify-content-center ">
+      <td className="w-25">
+        <div className="d-flex justify-content-center">
           <Link
             to={`/pacient/edit/${pacients?._id}`}
-            className="btn btn-warning my-2"
+            className= "btn btn-warning mx-1 "
           >
-            Update
+            Update 
           </Link>
           <button
-            className="btn btn-danger my-2"
+            className="btn btn-danger mx-1"
             onClick={() => handleDelete(pacients?._id)}
           >
             Delete
@@ -63,7 +64,6 @@ const Pacients = ({ pacients, getAPI }) => {
         </div>
       </td>
     </tr>
- 
   );
 };
